+++ CHANGELOG.rst
@@ -30,6 +30,8 @@ Fixed
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
 - Eventual failover may miss an event while roles are being reconfigured.
+- Compatibility with pipe logging
+  (see: https://github.com/tarantool/tarantool/issues/5220)
 
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 Enhanced is WebUI